+++ charts/dapr/charts/dapr_operator/templates/dapr_operator_deployment.yaml
@@ -172,12 +172,12 @@ spec:
 {{- if .Values.global.operator.watchdogCanPatchPodLabels }}
         - "--watchdog-can-patch-pod-labels"
 {{- end }}
-{{- if .Values.global.actors.enabled }}
-        - "--additional-control-plane-services"
+{{- if .Values.global.actors.serviceName }}
+        - "--additional-control-plane-service"
         - "{{ .Values.global.actors.serviceName }}"
 {{- end }}
 {{- if .Values.global.reminders.serviceName }}
-        - "--additional-control-plane-services"
+        - "--additional-control-plane-service"
         - "{{ .Values.global.reminders.serviceName }}"
 {{- end }}
       serviceAccountName: dapr-operator
