+++ cmd/sentry/options/options.go
@@ -78,7 +78,7 @@ func New(origArgs []string) *Options {
 	fs.StringVar(&opts.TrustDomain, "trust-domain", "localhost", "The CA trust domain")
 	fs.IntVar(&opts.Port, "port", config.DefaultPort, "The port for the sentry server to listen on")
 	fs.IntVar(&opts.HealthzPort, "healthz-port", 8080, "The port for the healthz server to listen on")
-	fs.StringSliceVar(&opts.AdditionalCPServices, "additional-control-plane-services", []string{"placement"}, "Name of the additional control plane services, if any")
+	fs.StringSliceVar(&opts.AdditionalCPServices, "additional-control-plane-service", []string{"placement"}, "Name of the additional control plane services, if any")
 
 	if home := homedir.HomeDir(); home != "" {
 		fs.StringVar(&opts.Kubeconfig, "kubeconfig", filepath.Join(home, ".kube", "config"), "(optional) absolute path to the kubeconfig file")
